refactor: folder api 및 hook 통합
