+++ src/hooks/folder/use-folder.ts
@@ -1,6 +1,20 @@
+import { deleteFolder } from "@apis/folder/folder";
 import { FOLDER_QUERY_KEY } from "@apis/folder/folder-queries";
-import { toggleFolderMark } from "@apis/folder/mark-folder";
+import { toggleFolderMark } from "@apis/folder/folder";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
+
+export const useDeleteFolderMutation = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: (folderId: number) => deleteFolder(folderId),
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: FOLDER_QUERY_KEY.ALL,
+      });
+    },
+  });
+};
 
 export const useFolderMark = () => {
   const queryClient = useQueryClient();
